Staff page is working again
- Adjust naming for JSON files match with this project's naming
  convention.

- Also update file naming convention (Part 3)
